Update gradle plugin: Deprecate experimental configuration for Compose for Web
Since 1.6.10 Compose for Web goes to Alpha and experimental configuration is not needed anymore. We'll do it by default.
