+++ gradle-plugins/compose/src/main/kotlin/org/jetbrains/compose/experimental/dsl/ExperimentalExtension.kt
@@ -12,7 +12,15 @@ import javax.inject.Inject
 abstract class ExperimentalExtension @Inject constructor(
     objects: ObjectFactory
 ) {
+
+    @Deprecated(
+        message = "Starting from 1.6.10, Compose for Web goes to Alpha. Experimental configuration is not needed anymore.",
+    )
     val web: ExperimentalWebExtension = objects.newInstance(ExperimentalWebExtension::class.java)
+
+    @Deprecated(
+        message = "Starting from 1.6.10, Compose for Web goes to Alpha. Experimental configuration is not needed anymore."
+    )
     fun web(action: Action<ExperimentalWebExtension>) {
         action.execute(web)
     }
