+++ gradle-plugins/compose/src/main/kotlin/org/jetbrains/compose/experimental/dsl/ExperimentalWebApplication.kt
@@ -7,6 +7,9 @@ package org.jetbrains.compose.experimental.dsl
 
 import javax.inject.Inject
 
+@Deprecated(
+    message = "Starting from 1.6.10, Compose for Web goes to Alpha. Experimental configuration is not needed anymore.",
+)
 abstract class ExperimentalWebApplication  @Inject constructor(
     @Suppress("unused")
     val name: String,
