+++ gradle-plugins/compose/src/main/kotlin/org/jetbrains/compose/web/dsl/WebApplication.kt
@@ -1,0 +1,10 @@
+package org.jetbrains.compose.web.dsl
+
+import javax.inject.Inject
+
+abstract class WebApplication  @Inject constructor(
+    @Suppress("unused")
+    val name: String,
+) {
+
+}
